+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://git.rustybever.be/vieter-v/vieter/src/branch/dev)
 
+### Added
+
 * Metrics endpoint for Prometheus integration
+* Search in list of targets using API & CLI
 
 ## [0.5.0](https://git.rustybever.be/vieter-v/vieter/src/tag/0.5.0)